Update ShippingFile - It is starting to work
With that database scheme, I could progress in this object,
which might be moved into lib/objects, and replaced by the
Controller.

To advance, there were some edits in the Assignor and the
Cnab240File. The last one may be removed, since it would
not be of much use to the future ReturnFile.

Also, to avoid problems with accents, a new dependency was
introduced. NoDiacritic removes accents from characters.
And aryelgois\utils was updated.

Now, the FileHeader, LotHeader, LotTrailer and FileTrailer
seem to be working. Just LotDetail is missing, my next
target.