Work in progress
It is almost usable, what i can see that is missing is the
calculation of check digits and the writing of docblocks.

Also, I need a real world test, which will happen only after
bringing the bank billet generator.

This time, the database scheme got updated, and a lot of
things changed. The main change was the new titles table,
replacing services. So, the Service object was replaced
by the new Title object, which also contains the Payer.

The Controller was updated to use this new logic, as well
the ShippingFile. And this guy is now producing the output
correctly!

Other objects and composer.lock were also updated.